Revert "completes integration tests"
This reverts commit 3a8c26924f7531b7b3fc50f3ae7c2f3fbd8779cd.

+++ contracts/common/lib/ECVerify.sol
@@ -1,7 +1,12 @@
 pragma solidity ^0.5.2;
 
+
 library ECVerify {
-    function ecrecovery(bytes32 hash, uint256[3] memory sig) internal pure returns (address) {
+    function ecrecovery(bytes32 hash, uint[3] memory sig)
+        internal
+        pure
+        returns (address)
+    {
         bytes32 r;
         bytes32 s;
         uint8 v;
@@ -34,7 +39,11 @@ library ECVerify {
         return result;
     }
 
-    function ecrecovery(bytes32 hash, bytes memory sig) internal pure returns (address) {
+    function ecrecovery(bytes32 hash, bytes memory sig)
+        internal
+        pure
+        returns (address)
+    {
         bytes32 r;
         bytes32 s;
         uint8 v;
@@ -67,7 +76,11 @@ library ECVerify {
         return result;
     }
 
-    function ecrecovery(bytes32 hash, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
+    function ecrecovery(bytes32 hash, uint8 v, bytes32 r, bytes32 s)
+        internal
+        pure
+        returns (address)
+    {
         // get address out of hash and signature
         address result = ecrecover(hash, v, r, s);
 
@@ -77,7 +90,11 @@ library ECVerify {
         return result;
     }
 
-    function ecverify(bytes32 hash, bytes memory sig, address signer) internal pure returns (bool) {
+    function ecverify(bytes32 hash, bytes memory sig, address signer)
+        internal
+        pure
+        returns (bool)
+    {
         return signer == ecrecovery(hash, sig);
     }
 }
